Add support for using client certificates in JSVM
JSVM object now has reference to API Spec, and its upstream certificates

In addition getUpsteramCertificates function now handle if spec is nil
and apply global rules
