+++ reverse_proxy.go
@@ -31,7 +31,6 @@ import (
 	cache "github.com/pmylund/go-cache"
 
 	"github.com/TykTechnologies/tyk/apidef"
-	"github.com/TykTechnologies/tyk/certs"
 	"github.com/TykTechnologies/tyk/config"
 	"github.com/TykTechnologies/tyk/user"
 )
@@ -432,36 +431,6 @@ func httpTransport(timeOut int, rw http.ResponseWriter, req *http.Request, p *Re
 	}
 
 	return transport
-}
-
-func getUpstreamCertificate(host string, spec *APISpec) (cert *tls.Certificate) {
-	var certID string
-
-	for _, m := range []map[string]string{config.Global.Security.Certificates.Upstream, spec.UpstreamCertificates} {
-		if len(m) == 0 {
-			continue
-		}
-
-		if id, ok := m["*"]; ok {
-			certID = id
-		}
-
-		if id, ok := m[host]; ok {
-			certID = id
-		}
-	}
-
-	if certID == "" {
-		return nil
-	}
-
-	certs := CertificateManager.List([]string{certID}, certs.CertificatePrivate)
-
-	if len(certs) == 0 {
-		return nil
-	}
-
-	return certs[0]
 }
 
 func (p *ReverseProxy) WrappedServeHTTP(rw http.ResponseWriter, req *http.Request, withCache bool) *http.Response {
